api/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
docopt==0.6.2
fuzzywuzzy==0.18.0
gunicorn==20.1.0
idna==3.4
Levenshtein==0.20.8
Markdown==3.4.1
packaging==22.0
Pillow==9.3.0
pipreqs==0.4.11
pytesseract==0.3.10
python-Levenshtein==0.20.8
pytz==2022.7
rapidfuzz==2.13.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
yarg==0.1.9
